semantic_search_service/requirements.txt
pyzmq
sentence-transformers
numpy
sklearn
